id 타입은 Integer -> Long으로 변경

+++ back/src/main/java/com/back/domain/order/order/dto/OrderResponseDto.java
@@ -2,5 +2,5 @@ package com.back.domain.order.order.dto;
 
 import java.time.LocalDateTime;
 
-public record OrderResponseDto(Integer orderId, String ordererEmail, LocalDateTime orderedTime) {
+public record OrderResponseDto(Long orderId, String ordererEmail, LocalDateTime orderedTime) {
 }
